checkpoint: diagnostics module
