+++ src/context/mod.rs
@@ -1,26 +1,18 @@
 //! Structs that carry important global compiler state.
 
 use std::path::PathBuf;
-use std::cell::RefCell;
-use std::rc::Rc;
-use std::fmt;
-use std::any::Any;
 
 use crate::ident::GensymContext;
-use crate::error::CompileError;
 use crate::resolve::Resolutions;
 use crate::resolve::rib::Rib;
-
-use codespan_reporting as cs;
-use cs::term::termcolor as tc;
-
-use crate::pos::Files;
 
 pub use defs::Defs;
 pub mod defs;
 
 pub use consts::Consts;
 pub mod consts;
+
+pub use crate::diagnostic::DiagnosticEmitter;
 
 /// Context object for the majority of compilation.
 ///
@@ -76,20 +68,10 @@ pub struct CompilerContext<'ctx> {
     _lifetime: std::marker::PhantomData<*mut &'ctx ()>, // invariant
 }
 
-pub trait WriteError: tc::WriteColor + Any {
-    fn as_any(&self) -> &dyn Any;
-    fn as_write_color(&mut self) -> &mut dyn tc::WriteColor;
-}
-
-impl<T: tc::WriteColor + Any> WriteError for T {
-    fn as_any(&self) -> &dyn Any { self }
-    fn as_write_color(&mut self) -> &mut dyn tc::WriteColor { self }
-}
-
 impl CompilerContext<'_> {
-    fn from_diagnostic_writer<W: WriteError>(writer: W) -> Self {
+    fn from_diagnostic_emitter(diagnostics: DiagnosticEmitter) -> Self {
         CompilerContext {
-            diagnostics: DiagnosticEmitter::from_writer(writer),
+            diagnostics,
             mapfiles: Default::default(),
             resolutions: Default::default(),
             defs: Default::default(),
@@ -101,70 +83,9 @@ impl CompilerContext<'_> {
     }
 
     /// Create a [`CompilerContext`] that writes diagnostics to the standard error stream.
-    pub fn new_stderr() -> Self {
-        Self::from_diagnostic_writer(tc::StandardStream::stderr(tc::ColorChoice::Auto))
-    }
+    pub fn new_stderr() -> Self { Self::from_diagnostic_emitter(DiagnosticEmitter::new_stderr()) }
 
     /// Create a [`CompilerContext`] that captures diagnostic output which can be recovered
-    /// by calling [`Self::get_captured_diagnostics`].
-    pub fn new_captured() -> Self {
-        let writer: CapturedWriter = tc::NoColor::new(vec![]);
-        Self::from_diagnostic_writer(writer)
-    }
-
-    /// Obtain captured diagnostics written to stderr, provided that this [`CompilerContext`]
-    /// was constructed using [`Self::new_captured`]. (otherwise, returns `None`)
-    pub fn get_captured_diagnostics(&self) -> Option<String> {
-        let writer = self.diagnostics.writer.borrow().as_any().downcast_ref::<CapturedWriter>()?;
-
-        Some(String::from_utf8_lossy(&writer.get_ref()).into_owned())
-    }
-}
-
-type CapturedWriter = tc::NoColor<Vec<u8>>;
-
-// =============================================================================
-
-/// Type responsible for emitting diagnostics and storing the metadata necessary to render them.
-#[derive(Clone)]
-pub struct DiagnosticEmitter {
-    files: Files,
-    config: cs::term::Config,
-    writer: Rc<RefCell<dyn WriteError>>,
-}
-
-impl fmt::Debug for DiagnosticEmitter {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        f.debug_struct("DiagnosticEmitter")
-            .field("files", &self.files)
-            .field("config", &self.config)
-            .field("writer", &(..))
-            .finish()
-    }
-}
-
-impl DiagnosticEmitter {
-    fn from_writer<W: WriteError>(writer: W) -> Self {
-        DiagnosticEmitter {
-            files: Files::new(),
-            config: default_term_config(),
-            writer: Rc::new(RefCell::new(writer)),
-        }
-    }
-
-    pub fn emit(&self, error: CompileError) {
-        let mut writer = self.writer.borrow_mut();
-
-        error.emit_to_writer(writer.as_write_color(), &self.files, &self.config);
-    }
-}
-
-fn default_term_config() -> cs::term::Config {
-    let mut config = cs::term::Config::default();
-    // Make output closer to rustc. Fewer colors overall, looks better.
-    config.styles.primary_label_error.set_intense(true);
-    config.styles.secondary_label.set_intense(true);
-    config.styles.line_number.set_intense(true);
-    config.styles.source_border.set_intense(true);
-    config
+    /// by calling [`DiagnosticEmitter::get_captured_diagnostics`].
+    pub fn new_captured() -> Self { Self::from_diagnostic_emitter(DiagnosticEmitter::new_captured()) }
 }
